feat: import with interface ok 4 synthese
