+++ backend/geonature/core/gn_commons/models/__init__.py
@@ -1,2 +1,3 @@
 from .base import *
+from .abstract_interface import AbstractInterface
 from .additional_fields import *
