+++ backend/geonature/core/gn_commons/models/abstract_interface.py
@@ -1,0 +1,13 @@
+from .base import TModules
+import abc
+
+
+class AbstractInterface(abc.ABC):
+    _module: TModules = None
+
+    def __init__(self, module: TModules) -> None:
+        self._module = module
+
+    @classmethod
+    def is_implemented_in_module(cls, module: TModules) -> bool:
+        return isinstance(module, cls)
