+++ backend/geonature/core/imports/routes/imports.py
@@ -46,7 +46,6 @@ from geonature.core.imports.utils import (
     clean_import,
     generate_pdf_from_template,
 )
-from geonature.core.imports.bbox import get_valid_bbox
 from geonature.core.imports.tasks import do_import_checks, do_import_in_destination
 
 IMPORTS_PER_PAGE = 15
@@ -466,7 +465,7 @@ def preview_valid_data(scope, imprt):
         raise Conflict("Import must have been prepared before executing this action.")
 
     data = {
-        "valid_bbox": get_valid_bbox(imprt),
+        "valid_bbox": imprt.destination.interface_import.compute_bounding_box(imprt),
         "entities": [],
     }
 
@@ -641,7 +640,7 @@ def delete_import(scope, imprt):
     db.session.execute(
         delete(transient_table).where(transient_table.c.id_import == imprt.id_import)
     )
-    imprt.destination.remove_data_from_destination(imprt)
+    imprt.destination.interface_import.remove_data_from_destination(imprt)
     db.session.delete(imprt)
     db.session.commit()
     return jsonify()
@@ -713,4 +712,4 @@ def get_foreign_key_attr(obj, field: str):
 @permissions.check_cruved_scope("R", get_scope=True, module_code="IMPORT", object_code="IMPORT")
 def report_plot(scope, imprt: TImports):
 
-    return json.dumps(imprt.destination.plot_function(imprt))
+    return json.dumps(imprt.destination.interface_import.report_plot(imprt))
