+++ backend/geonature/core/imports/tasks.py
@@ -32,7 +32,7 @@ def do_import_checks(self, import_id):
     check_orphan_rows(imprt)
     self.update_state(state="PROGRESS", meta={"progress": 0.1})
 
-    imprt.destination.check_transient_data(self, logger, imprt)
+    imprt.destination.interface_import.check_transient_data(self, logger, imprt)
 
     self.update_state(state="PROGRESS", meta={"progress": 1})
 
@@ -68,7 +68,7 @@ def do_import_in_destination(self, import_id):
     transient_table = imprt.destination.get_transient_table()
 
     # Copy valid transient data to destination
-    imprt.destination.import_data_to_destination(imprt)
+    imprt.destination.interface_import.import_data_to_destination(imprt)
 
     imprt.import_count = db.session.execute(
         db.select(func.count())
